+++ src/Data/userData.ts
@@ -1,48 +1,33 @@
+import { User } from "../Models/User";
 import { BaseDataBase } from "./BaseDatabase";
 
-export default class BandCreatedData extends BaseDataBase {
-    protected TABLE_NAME = 'Bandas'
 
-    insert = async (band: any) => {
-        try {
-            const response = await this.connection(this.TABLE_NAME)
-                .insert(band)
-            return response
-        } catch (error) {
-            if (error instanceof Error) {
-                throw new Error(error.message)
-            } else {
-                throw new Error('Erro')
-            }
-
-        }
+export class UserData extends BaseDataBase {
+    signup = async (user:User) => {
+        await this.connection("Users_music")
+        .insert({
+            id: user.id,
+            name: user.name,
+            email: user.email,
+            password: user.password,
+            role: user.role
+        })
     }
 
-    findByBand = async (name: string) => {
+    selectUserByEmail = async (email:string) => {
+
         try {
-            const queryResult = await this.connection(this.TABLE_NAME)
-                .select()
-                .where({ name })
-            return queryResult[0]
-        } catch (error) {
-            throw new Error()
+
+        const response = await this.connection("Users_music")
+        .where({email:email})
+
+        return response[0]
+
+        } catch (error:any) {
+           throw new Error(error.message);
 
         }
+
     }
-
-
-
-    //     findByPassword = async (password: string) => {
-    //         try {
-    //             const queryResult = await BaseDataBase
-    //                 .connection(this.TABLE_NAME)
-    //                 .select()
-    //                 .where({ password })
-    //             return queryResult[0]
-    //         } catch (error) {
-    //             throw new Error()
-
-    //         }
-    //     }
 
 }
